Remove mark_used function to reduce binary size
The mark_used function was never called, it existed so the contract
could have a reference to the `pay_for_memory_grow` host-io. We moved
this reference to the entrypoint, saving 360 bytes of the final
compressed binary size. The reason why we need the call to
`pay_for_memory_grow` is explained in the code.